fix(Scanner)!: Fail early on misconfigured version and criteria
If the criteria for looking up scan results do no match the scanner
details of the available scanner, no scan results won't be looked-up
from the configured storage.

In case the storage has scan results matching the details of the
availble scanner the `Scanner` will re-scan without need instead of
re-using the scan results from the storage. It will also attempt to
overwrite the already existing scan results in the storage which will
result in contraint violation exceptions in case of a `PostgresStorage`.

As this kind of mis-configuration causes large amounts of unnecessary
work in the following way, it better to not tolerate that
mis-configuraiton. It makes it harder to configure ORT in a wrong way.

1. users needing to investigate what went wrong
2. CI runs consuming computation time due to un-cached scans

Fixes #6581.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>