feat(cli): Use the resolved resolutions in the reporter command
If the `ResolvedConfiguration` contains resolutions, use only those
resolutions in the `ReporterCommand`. This ensures that the used
resolutions are consistent with those used by the evaluator.

If the `ResolvedConfiguration` does not contain resolutions, read the
resolutions from the default sources as before. This way the reporter
can still be used without running the evaluator before.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>